PR updates

+++ src/subscribe_handler.cc
@@ -11,7 +11,7 @@
 
 namespace laps {
     SubscribeTrackHandler::SubscribeTrackHandler(const quicr::FullTrackName& full_track_name,
-                                                 std::uint64_t priority,
+                                                 std::uint8_t priority,
                                                  std::optional<quicr::messages::GroupOrder> group_order,
                                                  ClientManager& server,
                                                  std::weak_ptr<timeq::tick_service> tick_service,
